Reducing some warnings
we're going purely 64bit across all platforms, reducing the amounts of spurious 'ints' laying around where a long long would be perfectly fine (no risk of overflow since long long can hold anything an int can)

+++ p_sight.c
@@ -217,7 +217,7 @@ boolean P_CheckSight(mobj_t *t1, mobj_t *t2)
 {
   const sector_t *s1 = t1->subsector->sector;
   const sector_t *s2 = t2->subsector->sector;
-  int pnum = (s1-sectors)*numsectors + (s2-sectors);
+  long long pnum = (s1-sectors)*numsectors + (s2-sectors);
   los_t los;
 
   // First check for trivial rejection.
